polish-8: RUF lint blocking + fees router + bench 130->145
- Promoted ruff RUF (Ruff-specific lints) to blocking. `ruff --fix`
  cleaned 200 sites in one pass — mostly RUF100 unused-noqa, RUF102
  invalid-rule-code, RUF021 parenthesize-chained-operators.
  `--unsafe-fixes --fix` cleaned 30 more: RUF005 list-concat
  (`a + [x]` -> `[*a, x]`), RUF046 unnecessary cast-to-int
  (`int(round(x))` -> `round(x)`), RUF034 useless if-else.

  The 104 unicode-ambiguous findings (RUF001/2/3 — em-dashes,
  smart quotes, Devanagari in Indian-language UI strings +
  golden-bench answers like "मुख्यमंत्री") are codebase-ignored
  as intentional, not typos. The remaining 18 carry targeted
  `# noqa: RUFxxx`/`# noqa: SIMxxx` markers — 4 RUF012 mutable-
  class-default on read-only lookup dicts (`_VOICE_MAP` in tts.py
  + test fixtures), 14 SIM revisits that came in after the
  polish-6 SIM cleanup. Future regressions caught at PR time.

- Tenth router slice: 7 `/api/orgs/{id}/fees*` endpoints
  (structures + invoices + Razorpay flow) moved to
  `padhai/routers/orgs_fees.py`. Lifted the two marshalling
  helpers (_fee_struct_to_dict, _invoice_to_dict) with them as
  this was their only call site. ~190 lines off web.py. The
  Razorpay webhook stays in web.py — it dispatches BOTH fees
  + subscription-tier events.

- Accuracy bench 130 -> 145 items:
  * 4 more hard: NEET current-electricity (wire stretched), JEE
    limits (sin x / x), UPSC Sixth Schedule, JEE hybridization.
  * 4 CBSE Class 6-8 items (was sparse): planet Mercury, 7^0,
    annual plants, helium for weather balloons.
  * 3 chemistry-depth: pH neutral=7, fluorine most electronegative,
    propanone IUPAC name.
  * 2 IGCSE items — entirely new board coverage (newton SI unit
    of force, chlorophyll).
  * 2 SSC fillers: Ambedkar Father of Constitution, longest
    river is Ganga.

  Distribution: easy=94 / medium=36 / hard=15.
  Structural runner: 145/145 in 3.1s.

CLAUDE.md §16 updated.

Verification:
- ruff F+E+I+B+UP+SIM+RUF -> clean
- pytest tests/ -> 37 passed
- 6 fees routes registered on app boot (7 endpoints; 1 shared URL)
- bench structural: 145/145

+++ padhai/affiliates.py
@@ -433,7 +433,7 @@ def book_commission(
     aff = get_affiliate_by_code(attribution.affiliate_code)
     if not aff or aff.status != "active":
         return None
-    commission = int(round(invoice_paise * aff.commission_pct))
+    commission = round(invoice_paise * aff.commission_pct)
     eid = uuid.uuid4().hex
     with _conn() as conn:
         try:
